Optimize constructors allocation for eye, ones, Inf, NaN, rand, randn functions.

+++ modules/elementary_functions/builtin/cpp/numelBuiltin.cpp
@@ -34,7 +34,7 @@ ArrayOfVector Nelson::ElementaryFunctionsGateway::numelBuiltin(Evaluator* eval, 
         Error(eval, ERROR_WRONG_NUMBERS_OUTPUT_ARGS);
     }
     ArrayOf param1 = argIn[0];
-    retval.push_back(ArrayOf::doubleConstructor(param1.numel()));
+    retval.push_back(ArrayOf::doubleConstructor((double)param1.numel()));
     return retval;
 }
 //=============================================================================
